+++ tests/Controller/Backend/AdminUserControllerTest.php
@@ -61,6 +61,8 @@ class AdminUserControllerTest extends WebTestCase
             ['/admin/app/archive/list'],
             ['/admin/app/archive/create'],
             ['/admin/app/archive/1/edit'],
+            ['/admin/app/configfooterinformation/list'],
+            ['/admin/app/configfooterinformation/1/edit'],
         ];
     }
 
@@ -97,6 +99,10 @@ class AdminUserControllerTest extends WebTestCase
             ['/admin/app/archive/9/edit'],
             ['/admin/app/archive/1/show'],
             ['/admin/app/archive/1/delete'],
+            ['/admin/app/configfooterinformation/create'],
+            ['/admin/app/configfooterinformation/9/edit'],
+            ['/admin/app/configfooterinformation/1/show'],
+            ['/admin/app/configfooterinformation/1/delete'],
             ['/admin/app/user/9/edit'],
             ['/admin/app/user/1/show'],
             ['/admin/app/user/1/delete'],