+++ tests/Controller/Backend/SuperAdminUserControllerTest.php
@@ -64,6 +64,8 @@ class SuperAdminUserControllerTest extends WebTestCase
             ['/admin/app/user/list'],
             ['/admin/app/user/create'],
             ['/admin/app/user/1/edit'],
+            ['/admin/app/configfooterinformation/list'],
+            ['/admin/app/configfooterinformation/1/edit'],
         ];
     }
 
@@ -100,6 +102,10 @@ class SuperAdminUserControllerTest extends WebTestCase
             ['/admin/app/archive/9/edit'],
             ['/admin/app/archive/1/show'],
             ['/admin/app/archive/1/delete'],
+            ['/admin/app/configfooterinformation/create'],
+            ['/admin/app/configfooterinformation/9/edit'],
+            ['/admin/app/configfooterinformation/1/show'],
+            ['/admin/app/configfooterinformation/1/delete'],
             ['/admin/app/user/9/edit'],
             ['/admin/app/user/1/show'],
             ['/admin/app/user/1/delete'],